style: api interceptor 위치를 api 정의한 곳으로 이동
